DeleteCollection,FaceAuthentication,FolderCheck,ListCollection,SearchFaceMatchImageCollection,Welcome,SaveImage
DeleteCollection 추가 - collection 삭제기능
FaceAuthentication 수정 - switch문 진입시 외부 반복문 탈출을 위한 check인덱스 추가
FolderCheck 수정 - collection 삭제시 하위폴더 삭제를 위해 조건문내에 재귀호출 추가
ListCollection 수정 - AwsVo에 collectionName 추가, 매개변수로 받은 email의
collection만 조회하도록 if문 추가
SearchFaceMatchImageCollection 수정 - withFaceMatchThreshold(70F->60F) 값변경
Welcome 작성 - 처음 가입했을시 실행되어야할 클래스, 기본 폴더 및 collection 생성
SaveImage 수정 - url을 받는 생성자에서 collection 매개변수 제거
controller 수정 - 위의 수정사항에 맞게 수정

+++ com/FolderCheck.java
@@ -55,17 +55,22 @@ public class FolderCheck {
 				File[] folder_list = folder.listFiles(); // 파일리스트 얻어오기
 
 				for (int j = 0; j < folder_list.length; j++) {
-					folder_list[j].delete(); // 파일 삭제
+					if(folder_list[j].isFile()){
+					folder_list[j].delete();
 					System.out.println(collectionName+"폴더내 파일이 삭제되었습니다.");
-
-				}
-
+					}else {
+						FolderDelete(email, collectionName+File.separator+folder_list[j].getName());
+						//재귀호출로 해당 폴더 경로에서 delete메서드 다시 실행
+						}
+					}
 				if (folder_list.length == 0 && folder.isDirectory()) {
 					folder.delete(); // 대상폴더 삭제
 					System.out.println(collectionName+" 폴더가 삭제되었습니다.");
 				}
+				
+				}
 			}
-		} catch (Exception e) {
+		 catch (Exception e) {
 			e.getStackTrace();
 		}
 	}
